Move TType -> binary conversions out of top-level
