+++ lib/rs/src/protocol/binary.rs
@@ -17,9 +17,9 @@
 
 use byteorder::{BigEndian, ByteOrder, ReadBytesExt, WriteBytesExt};
 use std::cell::RefCell;
-use std::convert;
+use std::convert::From;
 use std::rc::Rc;
-use try_from;
+use try_from::TryFrom;
 
 use ::{ProtocolError, ProtocolErrorKind};
 use ::transport::TTransport;
@@ -39,7 +39,7 @@ pub struct TBinaryProtocol {
 
 impl TBinaryProtocol {
     fn write_transport(&mut self, buf: &[u8]) -> ::Result<()> {
-        self.transport.borrow_mut().write(buf).map(|_| ()).map_err(convert::From::from)
+        self.transport.borrow_mut().write(buf).map(|_| ()).map_err(From::from)
     }
 }
 
@@ -113,23 +113,23 @@ impl TProtocol for TBinaryProtocol {
     }
 
     fn write_i8(&mut self, i: i8) -> ::Result<()> {
-        self.transport.borrow_mut().write_i8(i).map_err(convert::From::from)
+        self.transport.borrow_mut().write_i8(i).map_err(From::from)
     }
 
     fn write_i16(&mut self, i: i16) -> ::Result<()> {
-        self.transport.borrow_mut().write_i16::<BigEndian>(i).map_err(convert::From::from)
+        self.transport.borrow_mut().write_i16::<BigEndian>(i).map_err(From::from)
     }
 
     fn write_i32(&mut self, i: i32) -> ::Result<()> {
-        self.transport.borrow_mut().write_i32::<BigEndian>(i).map_err(convert::From::from)
+        self.transport.borrow_mut().write_i32::<BigEndian>(i).map_err(From::from)
     }
 
     fn write_i64(&mut self, i: i64) -> ::Result<()> {
-        self.transport.borrow_mut().write_i64::<BigEndian>(i).map_err(convert::From::from)
+        self.transport.borrow_mut().write_i64::<BigEndian>(i).map_err(From::from)
     }
 
     fn write_double(&mut self, d: f64) -> ::Result<()> {
-        self.transport.borrow_mut().write_f64::<BigEndian>(d).map_err(convert::From::from)
+        self.transport.borrow_mut().write_f64::<BigEndian>(d).map_err(From::from)
     }
 
     fn write_string(&mut self, s: &str) -> ::Result<()> {
@@ -165,7 +165,7 @@ impl TProtocol for TBinaryProtocol {
     }
 
     fn flush(&mut self) -> ::Result<()> {
-        self.transport.borrow_mut().flush().map_err(convert::From::from)
+        self.transport.borrow_mut().flush().map_err(From::from)
     }
 
     fn read_message_begin(&mut self) -> ::Result<TMessageIdentifier> {
@@ -188,7 +188,7 @@ impl TProtocol for TBinaryProtocol {
                     )
                 )
             } else {
-                let message_type: TMessageType = try!(try_from::TryFrom::try_from(first_bytes[3]));
+                let message_type: TMessageType = try!(TryFrom::try_from(first_bytes[3]));
                 let name = try!(self.read_string());
                 let sequence_number = try!(self.read_i32());
                 Ok(TMessageIdentifier { name: name, message_type: message_type, sequence_number: sequence_number })
@@ -217,7 +217,7 @@ impl TProtocol for TBinaryProtocol {
                 let name = try!(String::from_utf8(name_buf));
 
                 // read the rest of the fields
-                let message_type: TMessageType = try!(self.read_byte().and_then(try_from::TryFrom::try_from));
+                let message_type: TMessageType = try!(self.read_byte().and_then(TryFrom::try_from));
                 let sequence_number = try!(self.read_i32());
                 Ok(TMessageIdentifier { name: name, message_type: message_type, sequence_number: sequence_number })
             }
@@ -238,7 +238,7 @@ impl TProtocol for TBinaryProtocol {
 
     fn read_field_begin(&mut self) -> ::Result<TFieldIdentifier> {
         let field_type_byte = try!(self.read_byte());
-        let field_type = try!(try_from::TryFrom::try_from(field_type_byte));
+        let field_type = try!(TryFrom::try_from(field_type_byte));
         let id = try!(match field_type {
             TType::Stop => Ok(0),
             _ => self.read_i16()
@@ -253,7 +253,7 @@ impl TProtocol for TBinaryProtocol {
     fn read_bytes(&mut self) -> ::Result<Vec<u8>> {
         let num_bytes = try!(self.transport.borrow_mut().read_i32::<BigEndian>()) as usize;
         let mut buf = vec![0u8; num_bytes];
-        self.transport.borrow_mut().read_exact(&mut buf).map(|_| buf).map_err(convert::From::from)
+        self.transport.borrow_mut().read_exact(&mut buf).map(|_| buf).map_err(From::from)
     }
 
     fn read_bool(&mut self) -> ::Result<bool> {
@@ -265,32 +265,32 @@ impl TProtocol for TBinaryProtocol {
     }
 
     fn read_i8(&mut self) -> ::Result<i8> {
-        self.transport.borrow_mut().read_i8().map_err(convert::From::from)
+        self.transport.borrow_mut().read_i8().map_err(From::from)
     }
 
     fn read_i16(&mut self) -> ::Result<i16> {
-        self.transport.borrow_mut().read_i16::<BigEndian>().map_err(convert::From::from)
+        self.transport.borrow_mut().read_i16::<BigEndian>().map_err(From::from)
     }
 
     fn read_i32(&mut self) -> ::Result<i32> {
-        self.transport.borrow_mut().read_i32::<BigEndian>().map_err(convert::From::from)
+        self.transport.borrow_mut().read_i32::<BigEndian>().map_err(From::from)
     }
 
     fn read_i64(&mut self) -> ::Result<i64> {
-        self.transport.borrow_mut().read_i64::<BigEndian>().map_err(convert::From::from)
+        self.transport.borrow_mut().read_i64::<BigEndian>().map_err(From::from)
     }
 
     fn read_double(&mut self) -> ::Result<f64> {
-        self.transport.borrow_mut().read_f64::<BigEndian>().map_err(convert::From::from)
+        self.transport.borrow_mut().read_f64::<BigEndian>().map_err(From::from)
     }
 
     fn read_string(&mut self) -> ::Result<String> {
         let bytes = try!(self.read_bytes());
-        String::from_utf8(bytes).map_err(convert::From::from)
+        String::from_utf8(bytes).map_err(From::from)
     }
 
     fn read_list_begin(&mut self) -> ::Result<TListIdentifier> {
-        let element_type: TType = try!(self.read_byte().and_then(try_from::TryFrom::try_from));
+        let element_type: TType = try!(self.read_byte().and_then(TryFrom::try_from));
         let size = try!(self.read_i32());
         let ret = TListIdentifier { element_type: element_type, size: size };
         Ok(ret)
@@ -301,7 +301,7 @@ impl TProtocol for TBinaryProtocol {
     }
 
     fn read_set_begin(&mut self) -> ::Result<TSetIdentifier> {
-        let element_type: TType = try!(self.read_byte().and_then(try_from::TryFrom::try_from));
+        let element_type: TType = try!(self.read_byte().and_then(TryFrom::try_from));
         let size = try!(self.read_i32());
         let ret = TSetIdentifier { element_type: element_type, size: size };
         Ok(ret)
@@ -312,8 +312,8 @@ impl TProtocol for TBinaryProtocol {
     }
 
     fn read_map_begin(&mut self) -> ::Result<TMapIdentifier> {
-        let key_type: TType = try!(self.read_byte().and_then(try_from::TryFrom::try_from));
-        let value_type: TType = try!(self.read_byte().and_then(try_from::TryFrom::try_from));
+        let key_type: TType = try!(self.read_byte().and_then(TryFrom::try_from));
+        let value_type: TType = try!(self.read_byte().and_then(TryFrom::try_from));
         let size = try!(self.read_i32());
         let ret = TMapIdentifier { key_type: key_type, value_type: value_type, size: size };
         Ok(ret)
@@ -328,11 +328,69 @@ impl TProtocol for TBinaryProtocol {
     //
 
     fn write_byte(&mut self, b: u8) -> ::Result<()> {
-        self.transport.borrow_mut().write_u8(b).map_err(convert::From::from)
+        self.transport.borrow_mut().write_u8(b).map_err(From::from)
     }
 
     fn read_byte(&mut self) -> ::Result<u8> {
-        self.transport.borrow_mut().read_u8().map_err(convert::From::from)
+        self.transport.borrow_mut().read_u8().map_err(From::from)
+    }
+}
+
+// Converts a Thrift field-type enum into its
+// byte representation for encoding into its serialized form.
+impl From<TType> for u8 {
+    fn from(field_type: TType) -> Self {
+        match field_type {
+            TType::Stop => 0x00,
+            TType::Void => 0x01,
+            TType::Bool => 0x02,
+            TType::I08 => 0x03, // equivalent to TType::Byte
+            TType::Double => 0x04,
+            TType::I16 => 0x06,
+            TType::I32 => 0x08,
+            TType::I64 => 0x0A,
+            TType::String => 0x0B,
+            TType::Utf7 => 0x0B,
+            TType::Struct => 0x0C,
+            TType::Map => 0x0D,
+            TType::Set => 0x0E,
+            TType::List => 0x0F,
+            TType::Utf8 => 0x10,
+            TType::Utf16 => 0x11,
+        }
+    }
+}
+
+// Converts the serialized representation of a
+// Thrift field type into its enum form.
+impl TryFrom<u8> for TType {
+    type Err = ::Error;
+    fn try_from(b: u8) -> ::Result<Self> {
+        match b {
+            0x00 => Ok(TType::Stop),
+            0x01 => Ok(TType::Void),
+            0x02 => Ok(TType::Bool),
+            0x03 => Ok(TType::I08), // Equivalent to TType::Byte
+            0x04 => Ok(TType::Double),
+            0x06 => Ok(TType::I16),
+            0x08 => Ok(TType::I32),
+            0x0A => Ok(TType::I64),
+            0x0B => Ok(TType::String), // technically, also a UTF7, but we'll treat it as string
+            0x0C => Ok(TType::Struct),
+            0x0D => Ok(TType::Map),
+            0x0E => Ok(TType::Set),
+            0x0F => Ok(TType::List),
+            0x10 => Ok(TType::Utf8),
+            0x11 => Ok(TType::Utf16),
+            unkn => Err(
+                ::Error::Protocol(
+                    ProtocolError {
+                        kind: ProtocolErrorKind::InvalidData,
+                        message: format!("cannot convert {} to TType", unkn),
+                    }
+                )
+            )
+        }
     }
 }
